Fix Loop/OnceBackwards tests
The tests were previously failing because the first (in this case last) keyframe was being shown twice. This sets the time to the max keyframe time, which is less than the duration. Some animations have clips with varying durations, so taking the longest will make sure we don't skip any frames.